src/requirements.txt
python-dotenv==1.0.1
Flask==3.0.2
Flask-Cors==4.0.0
Flask-PyMongo==2.3.0
gunicorn==20.1.0
requests==2.31.0
pynpm==0.2.0
json-with-comments==1.2.4
textract==1.6.5
